Prevent crash in j2objc transpiled code
J2objc uses jre_emul hashmap implementation.

A particular feature of a map entrySet method causes problems:
“Returns a Set view of the mappings contained in this map. The set is backed by the map, so changes to the map are reflected in the set, and vice-versa.”

The problems is caused by the iOS implementation that keeps the backing map as a @weakOuter. Because of this, the Set returned by the entrySet() method cannot safely outlive it’s backing Hashmap.

By copying the entrySet into a standard hashSet, we break the link, but allow the set to outlive it’s backing map.

Since the StreamApi is a functionnal API, it should not pose any problem