Don't flatten header directory to make it compatible with J2ObjC 2.4
Flattening the directory meant that it was not possible to differentiate the lightweight stream api classes from the java ones included in J2ObjC.

Previously, clashing files were renamed, but it introduced the need to have a header mapping in each project using this lib.

Instead, keep the directory intact.

The only annoying part about keeping the directory intact is that you must always reference to the complete path.

When manually typing the #include it's a pain, but, since there is no point in using this library directly in Objective-C/Swift (a native API exists), this pain does not exist.